+++ abbreviate.test.js
@@ -1,10 +1,10 @@
 'use strict';
 
-const abr = require('./abbreviate');
+const abbreviate = require('./index').abbreviate;
 
 function test_abbreviateStringByLength() {
   const expected =  {abbreviated: "AT", shortened: "an-t", rest: "u-tu", length: 4};
-  let out = abr.abbreviateStringByLength('an-tu-tu', 4, /[ -]/);
+  let out = abbreviate.stringByLength('an-tu-tu', 4, /[ -]/);
   console.log(out.abbreviated === expected.abbreviated);
   console.log(out.shortened === expected.shortened);
   console.log(out.rest === expected.rest);
@@ -12,21 +12,21 @@ function test_abbreviateStringByLength() {
 }
 
 function test_abbreviateString() {
-  const expected =  ["GFGF"];
-  let out = abr.abbreviateStrings(['git-faf Git Faf'], /[- ]/);
-  console.log(out[0] === expected[0]);
+  const expected =  "GFGF";
+  let out = abbreviate.aString('git-faf Git Faf', /[- ]/);
+  console.log(out === expected);
 }
 
 function test_abbreviateStrings() {
   const expected =  ["ATT", "USA"];
-  let out = abr.abbreviateStrings(['an-tu-tu', 'united states of america'], /[- of]/);
+  let out = abbreviate.manyStrings(['an-tu-tu', 'united states of america'], /[- of]/);
   console.log(out[0] === expected[0]);
   console.log(out[1] === expected[1]);
 }
 
 function test_uppercaseAutoAbbreviate () {
   const expected = "RI";
-  let out = abr.uppercaseAutoAbbreviate("Republic of India");
+  let out = abbreviate.autoUppercase("Republic of India");
   console.log(out === expected);
 }