+++ index.js
@@ -1,0 +1,5 @@
+const abbreviate = require('./abbreviate').abbreviate;
+
+module.exports = {
+    abbreviate
+};