+++ changelog.md
@@ -1,3 +1,9 @@
+# 2.0.0
+
+1. Introduces index.js
+2. Routes test file to use index.js instead of abbreviate.js
+3. All the tests pass. `npm test`.
+
 # 1.0.3
 
 Published on 16 August, 2017.